Build custom 'jar-server-demo' image when Docker credentials are provided
To be able to run the Java based demo, we run a static file server for the
Jar file. Although I previously build and uploaded an image to my own
Docker hub account, people should be able to build and push their own
automatically when `ANSIBLE_DOCKER_USERNAME` and `ANSIBLE_DOCKER_PASSWORD`
are provided as environment variables.