Fix race condition during deployment
The Nomad cluster with its three nodes must be up first, a quorum must have
been reached and a leader elected before we actually deploy workloads.

Ansible executes tasks for the same host in one go, even if tasks are
within a separate `hosts` section in the playbook. So the copying of the
job files and the execution were split in two parts.